building_practice/requirements.txt
langchain == 0.3.0
openai == 1.55.3
azure-search-documents == 11.6.0b2
pypdf == 4.3.1
streamlit == 1.37.1
python-dotenv == 1.0.1
Pillow>=10.2.0
pandas